requirements.txt
browser-cookie3==0.20.1
fastapi==0.135.3
gemini-webapi==1.21.0
httpx==0.28.1
notebooklm-cli==0.1.12
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.20
python-docx==1.2.0
python-dotenv==1.2.2
uvicorn==0.43.0
